requirements.txt
# Core audio processing
librosa>=0.10.0
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.10.0

# Speech-to-text
openai-whisper>=20231117

# Machine learning & embeddings
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.35.0

# Emotion detection
funasr>=1.0.0

# Acoustic features (eGeMAPS)
opensmile>=2.5.0

# Groq API
groq>=0.4.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
rich>=13.0.0
tqdm>=4.66.0

# Optional: for API server
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
